Update Testing Previsibine.ps1
Decided not to include a start where you exited off because that just over complicates the code.
Added more stuff to the DATA hash.
Renamed the settings text file to match the script.
Added support for the command line args in the script to some functions.
Rewrote the main function (might have to separate more things into their own functions...the main function is getting too big.
Corrected the variable to be saved in the Set-CK Function.
Fixed an issue regarding placeholders in the Set-CK and Get-ESPExtension functions.
Added an additional parameter to Set-xEdit to account for the settings file. More changes like this will come soon(tm).