Update Testing Previsibine.ps1
Don't really recall why I did the script parameters like that, but it was causing confusion when running Test-Path. Renamed function Set-xEdit to Set-XEdit. Added Write-File so all the writing to the settings file will be in that function. Fixed Test-CK and Test-XEdit with the boolean returns. Added a Test-FO4 function and removed some code from Test-CK. Get-CK just returns the CK. Set-CK was commented out because you cannot set the CK directory if the FO4 directory is defined (CK must be where FO4 is). Commented out the previous Set-XEdit code, split it into Get-XEdit, Set-XEdit, and Test-XEdit, will have to do the writing/overriding to settings file again.