Update Pwsh script
Removed the Xedit script parameter, I may set it again once I fix a few quirks. Added and sorted more messages to the hash table. Main function testing working as intended so far. Removed the Continue prompt for the Save-ESP1 function for now. Replaced the returns in Test-f4ck to Boolean for consistency with rest of script. Added a check to the Test-FO4CK function to read the file, removed variables in which I thought were pointless, and added more returns so the main function works as intended.  Edit-FO4CK removed some variables that I thought were pointless, moved the text replacement into the do-until loop. Test-CK's parameter setting has been removed., also removed more variables. Created Test-XEditFile, Set-XeditFile, Get-XEditFile, and Edit-XEditFile to mimic the functions that relate to FO4/CK. Added a Settings file (basically a file to be read from) so I don't have to do this limbo of creating it then deleting it again. Will delete when I am actually done with it.