Update Testing Previsibine.ps1
Added more messages to the hashtable. Added missing process functions to the main function. Renamed a parameter in Save-ESP1 and added missing FO4Edit processes to the CK choice. The Start-CK1 function now removes the Meshes folder once the archiving is done. Added checks to make sure there is an archive in the data folder. Made the Path variables for xedit and fo4 go to the new path in Edit-FO4CK and Edit-XEditFile.