Update Testing Previsibine.ps1
Alpha release.  Cleaned up Messages and comments throughout the mod that are not used or not planning to use.  Changed a few Message strings for Previs/Precombines.  Uncommented the functions in the main file. Completed the Start-Previs Function. Commented out the Disclaimer, I will add it once I feel like this script is a version 1.0. In Beta I will probably add the descriptive comments.